pyproject.toml
[tool.poetry]
name = "sample-spa"
version = "0.1.0"
description = ""
authors = ["Aleksandr Komissarov <alex.komissarov@ingrammicro.com>"]
readme = "README.md"
packages = [{include = "backend"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.99.1"
black = "^23.3.0"
uvicorn = "^0.22.0"
python-dotenv = "^1.0.0"
requests = "^2.31.0"
pytest = "^7.4.0"
pytest-asyncio = "^0.21.0"
pytest-mock = "^3.11.1"
isort = "^5.12.0"
connect-openapi-client = "^25.22"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
